mudança de Long para Integer

+++ sistema_petshop/src/main/java/com/petshop/model/Categoria.java
@@ -17,7 +17,8 @@ public class Categoria {
     // Declaração das variáveis
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
+    
     private String nome;
 
     @OneToMany(mappedBy = "categoria", cascade = CascadeType.ALL, orphanRemoval = true)
@@ -27,11 +28,11 @@ public class Categoria {
     public Categoria() {
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
 
@@ -43,7 +44,7 @@ public class Categoria {
         this.nome = nome;
     }
 
-    public Categoria(Long id, String nome) {
+    public Categoria(Integer id, String nome) {
         this.id = id;
         this.nome = nome;
     }
