+++ sistema_petshop/src/main/java/com/petshop/model/Especie.java
@@ -17,7 +17,8 @@ public class Especie {
     // Declaração das variáveis
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
+
     private String nome;
 
     @OneToMany(mappedBy = "especie", cascade = CascadeType.ALL, orphanRemoval = true)
@@ -32,11 +33,11 @@ public class Especie {
     }
 
     // getters and setters
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
