+++ sistema_petshop/src/main/java/com/petshop/model/Formas_de_Pagamento.java
@@ -17,7 +17,8 @@ public class Formas_de_Pagamento {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
+
     private String nome;
 
     @OneToMany(mappedBy = "formas_de_pagamento", cascade = CascadeType.ALL, orphanRemoval = true, fetch = FetchType.LAZY)
@@ -32,11 +33,11 @@ public class Formas_de_Pagamento {
     }
 
     // getters and setters
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
