+++ sistema_petshop/src/main/java/com/petshop/model/Produto.java
@@ -19,9 +19,12 @@ public class Produto {
     // Declaração das variáveis
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
+
     private String nome;
+
     private double preco;
+    
     private String fotoPath; // Caminho da imagem
 
     @ManyToOne
@@ -45,11 +48,11 @@ public class Produto {
 
     // Getters and Setters
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
