+++ sistema_petshop/src/main/java/com/petshop/repository/Itens_de_PedidosRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface Itens_de_PedidosRepository extends JpaRepository<Itens_de_Pedidos, Long> {
+public interface Itens_de_PedidosRepository extends JpaRepository<Itens_de_Pedidos, Integer> {
 }
